Separate PlanItem inner classes
No longer resulting in recursive frontend loops
