+++ src/repository/definition/cmmn/caseplan/eventlistenerdefinition.ts
@@ -1,4 +1,4 @@
-import { MilestoneEventListenerDefinition } from "./planitem";
+import MilestoneEventListenerDefinition from "./milestoneeventlistenerdefinition";
 
 export default class EventListenerDefinition extends MilestoneEventListenerDefinition {
 }
