+++ src/repository/definition/cmmn/caseplan/milestonedefinition.ts
@@ -1,5 +1,5 @@
 import { Element } from "../../../../util/xml";
-import { MilestoneEventListenerDefinition } from "./planitem";
+import MilestoneEventListenerDefinition from "./milestoneeventlistenerdefinition";
 
 export default class MilestoneDefinition extends MilestoneEventListenerDefinition {
     static get infix() {
