+++ src/repository/definition/cmmn/caseplan/planitem.ts
@@ -14,6 +14,7 @@ import PlanningTableDefinition from "./planning/planningtabledefinition";
 import StageDefinition from "./stagedefinition";
 import FourEyesDefinition from "./task/workflow/foureyesdefinition";
 import RendezVousDefinition from "./task/workflow/rendezvousdefinition";
+import TaskStageDefinition from "./taskstagedefinition";
 
 export default class PlanItem extends CMMNElementDefinition {
     private applicabilityRuleRefs: ReferenceSet<ApplicabilityRuleReference>;
@@ -197,60 +198,5 @@ export default class PlanItem extends CMMNElementDefinition {
      */
     get entryTransition(): string {
         throw new Error('This method must be implemented in ' + this.constructor.name);
-    }
-}
-
-/**
- * Simple helper class to re-use logic across stages and tasks
- */
-export class TaskStageDefinition extends PlanItem {
-    planningTable?: PlanningTableDefinition;
-    constructor(importNode: Element, caseDefinition: CaseDefinition, public parent: TaskStageDefinition | PlanningTableDefinition) {
-        super(importNode, caseDefinition, parent);
-        this.planningTable = this.parseElement('planningTable', PlanningTableDefinition);
-    }
-
-    getPlanningTable() {
-        if (!this.planningTable) {
-            this.planningTable = super.createDefinition(PlanningTableDefinition);
-        }
-        return this.planningTable;
-    }
-
-    get isTask() {
-        return false;
-    }
-
-    get isStage() {
-        return false;
-    }
-
-    get transitions() {
-        return ['complete', 'create', 'disable', 'enable', 'exit', 'fault', 'manualStart', 'parentResume', 'parentSuspend', 'reactivate', 'reenable', 'resume', 'start', 'suspend', 'terminate'];
-    }
-
-    get defaultTransition() {
-        return 'complete';
-    }
-
-    get entryTransition() {
-        return 'start';
-    }
-}
-
-/**
- * Simple helper class to re-use logic across milestones and event listeners
- */
-export class MilestoneEventListenerDefinition extends PlanItem {
-    get transitions(): string[] {
-        return ['occur', 'create', 'reactivate', 'resume', 'suspend', 'terminate'];
-    }
-
-    get defaultTransition() {
-        return 'occur';
-    }
-
-    get entryTransition() {
-        return 'occur';
     }
 }
