+++ src/repository/definition/cmmn/caseplan/stagedefinition.ts
@@ -1,11 +1,12 @@
 import XML, { Element } from "../../../../util/xml";
 import CaseDefinition from "../casedefinition";
 import MilestoneDefinition from "./milestonedefinition";
-import PlanItem, { TaskStageDefinition } from "./planitem";
+import PlanItem from "./planitem";
 import PlanningTableDefinition from "./planning/planningtabledefinition";
 import CaseTaskDefinition from "./task/casetaskdefinition";
 import HumanTaskDefinition from "./task/humantaskdefinition";
 import ProcessTaskDefinition from "./task/processtaskdefinition";
+import TaskStageDefinition from "./taskstagedefinition";
 import TimerEventDefinition from "./timereventdefinition";
 import UserEventDefinition from "./usereventdefinition";
